refactored key-points
