+++ app/controllers/products_controller.rb
@@ -7,7 +7,7 @@ class ProductsController < ApplicationController
   before_action :store_location, only: %i[index]
 
   def index
-    @products = @products.includes([product_image_attachment: :blob]).by_archived(params[:archived] || false).recently_updated.page(params[:page])
+    @products = @products.includes([product_image_attachment: :blob]).unarchived.recently_updated.page(params[:page])
   end
 
   def new; end
@@ -47,7 +47,7 @@ class ProductsController < ApplicationController
   end
 
   def toggle_archived
-    if @product.update_attribute(:archived, !@product.archived?)
+    if @product.update(archived: !@product.archived?)
       flash[:notice] = @product.archived ? 'Product Unpublished' : 'Product Published'
     else
       flash[:alert] = @product.errors.full_messages.join(', ')
