+++ app/models/concerns/sortable.rb
@@ -1,4 +1,4 @@
-module Sort
+module Sortable
   extend ActiveSupport::Concern
   included do
     scope :recently_updated, -> { order(updated_at: :desc) }
