+++ app/views/addresses/_form.html.erb
@@ -31,7 +31,7 @@
               </div>
               <div id="map" class="w-80 h-80"></div>
               <div class="actions flex items-center flex-row justify-center">
-                <%= form.submit "Submit", class: " mt-5 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded" %>
+                <%= form.submit "Submit", class: "mt-5 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded" %>
               </div>
             </div>
           </div>
