+++ app/views/layouts/application.html.erb
@@ -9,28 +9,15 @@
     <%= stylesheet_link_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
     <%= javascript_pack_tag 'application', 'data-turbolinks-track': 'reload' %>
     <%= javascript_pack_tag 'typewriter' %>
-    <%# <script>
-      var i = 0;
-      var txt = 'We offer a streamlined solution for users to sell their meticulously curated products obtained through web scraping. With our user-friendly interface, interested buyers can easily peruse and purchase these unique offerings.';
-      var speed = 50;
-
-      function typeWriter() {
-        if (i < txt.length) {
-          document.getElementById("home_description").innerHTML += txt.charAt(i);
-          i++;
-          setTimeout(typeWriter, speed);
-        }
-      }
-    </script> %>
   </head>
   <% if current_user %>
     <% current_user_id = current_user.id %>
   <% end %>
-  <div id = "user", data-user-id = <%= current_user_id %>></div>
+  <div id="user", data-user-id=<%= current_user_id %>></div>
   <body>
     <% if user_signed_in? %>
       <div class="fixed w-full top-0 z-10">
-        <%= render 'devise/shared/navbar' %>
+        <%= render "devise/shared/navbar" %>
       </div>
       <div class="mx-auto max-w-7xl px-2 sm:px-6 lg:px-8 relative flex h-16 bg-gray-800"></div>
     <% end %>
