+++ app/views/messages/new.html.erb
@@ -31,7 +31,7 @@
     <div class=" pt-2 flex fixed bottom-0 w-full p-4 bg-white rounded-lg w-100">
       <%= form_for [@order, Message.new], remote: true do |f| %>
         <div class="flex flex-row w-full">
-          <%= f.text_field :content, required: true, placeholder: "Enter your message", :maxlength=>"110", class: "flex-1 mr-2 border-2 border-gray-300 py-2 px-4 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-600 focus:border-transparent" %>
+          <%= f.text_field :content, required: true, placeholder: "Enter your message", maxlength: "110", class: "flex-1 mr-2 border-2 border-gray-300 py-2 px-4 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-600 focus:border-transparent" %>
           <%= f.submit "Send", class: "ml-3 bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-3 rounded-lg" %>
         </div>
       <% end %>
