+++ app/views/orders/show.html.erb
@@ -8,14 +8,10 @@
   <div class="flex flex-col items-start space-y-12 mx-10">
     <h4 class="font-semibold text-lg">Order Amount:</h4>
     <p class="text-gray-700 mb-6"><%= number_to_currency(@order.bid.price, unit: "Rs", format: "%u. %n") %></p>
-    <% unless current_user.buyer? %>
-      <h4 class="font-semibold text-lg">Buyer Name:</h4>
-      <p class="text-gray-700 mb-6"><%= @order.buyer.full_name %></p>
-    <% end %>
-    <% unless current_user.seller? %>
-      <h4 class="font-semibold text-lg">Seller Name:</h4>
-      <p class="text-gray-700 mb-6"><%= @order.seller.full_name %></p>
-    <% end %>
+    <h4 class="font-semibold text-lg">Seller Name:</h4>
+    <p class="text-gray-700 mb-6"><%= @order.seller.full_name %></p>
+    <h4 class="font-semibold text-lg">Buyer Name:</h4>
+    <p class="text-gray-700 mb-6"><%= @order.buyer.full_name %></p>
     <h4 class="font-semibold text-lg">Pickup Date & Time:</h4>
     <p class="text-gray-700 mb-6"><%= @order.pickup_time.strftime('%d %B %Y | %I:%M:%S %p') %></p>
   </div>
