+++ config/initializers/pagination.rb
@@ -1,0 +1,1 @@
+WillPaginate.per_page = 10
